requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1
chromadb==0.4.15
sentence-transformers==2.2.2
pysqlite3-binary

